correction datatypes et documentation silver emploi pivot + nvx modèles gold

+++ dbt_odis/models/gold/education_moyenne.sql
@@ -11,7 +11,7 @@ select
   end as codgeo,
   extract(year from cast(rentree_scolaire || '-01-01' as date)) as "YEAR",
   case
-    when nombre_total_classes != 0 then round(nombre_total_eleves / nombre_total_classes,2)
+    when nombre_total_classes != 0 then round(nombre_total_eleves::numeric / nombre_total_classes::numeric,2)
   else null
   end as "Nombre_Moyen_Eleves_Classe",
   nombre_total_eleves as "Nombre total d'élèves",
